backend/chat_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pymongo==4.8.0

# LangChain (let resolver pick latest compatible trio)
langchain-core>=0.3.16,<0.4.0
langchain-community>=0.3.3,<0.4.0
langchain-openai>=0.2.7,<0.3.0
